requirements.txt
streamlit
langchain
spotipy
python-dotenv